requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=1.4.53
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
PyYAML>=6.0.1
requests>=2.31.0
